+++ dp/_verify.py
@@ -6,25 +6,24 @@ from dp._index_converter import _indices_to_np_indices
 
 @staticmethod
 def verify_traceback_solution(arr, solution):
-    """
-    Verify that solution is a valid traceback solution of the DPArray.
+    """Verify that solution is a valid traceback solution of the DPArray.
 
     Args:
-        dp (DPArray): A DPArray that should be initialzed according to an
-        optimization problem. A traceback solution is not well defined
-        for non-optimization dynamic programming (i.e. does not use
-        max or min).
+        dp (DPArray): A DPArray that should be initialzed according 
+            to an optimization problem. A traceback solution is not
+            well defined for non-optimization dynamic programming
+            (i.e. does not use max or min).
 
-        solution (list of indices): S list of indices.
-        For 1D DPArrays, this should be a list of integers.
-        For 2D DPArrays, this should be a list of tuples.
+        solution (list of indices): A list of indices.
+            For 1D DPArrays, this should be a list of integers.
+            For 2D DPArrays, this should be a list of tuples.
         
     Return:
-        bool: False if the given solution is not correct
-        and True if it is correct. Empty solutions are considered correct.
-        Incomplete solutions (in which an unexplored predecessor still
-        exists) are incorrect. Solutions that contain the index of an
-        unitialized element are incorrect.
+        bool: False if the given solution is not correct and True
+            if it is correct. Empty solutions are considered correct.
+            Incomplete solutions (in which an unexplored predecessor
+            still exists) are incorrect. Solutions that contain the
+            index of an unitialized element are incorrect.
     """
     # Handle trivial case.
     if len(solution) == 0:
@@ -36,17 +35,17 @@ def verify_traceback_solution(arr, solution):
 
     # Go through time steps and check predecessors.
     i = len(solution) - 1
+    name = arr.array_name
     for timestep in reversed(arr.get_timesteps()):
         # If writing solution[i], analyze predecessors.
-        if solution[i] in timestep[arr.array_name][Op.WRITE]:
+        if solution[i] in timestep[name][Op.WRITE]:
             # Check that solution[0] has no predecessors.
             if i == 0:
-                return len(timestep[arr.array_name][Op.HIGHLIGHT]) == 0
+                return len(timestep[name][Op.HIGHLIGHT]) == 0
 
             # solution[i - 1] is not a predecessor.
             # The given solution is not correct.
-            if solution[i - 1] not in \
-                    timestep[arr.array_name][Op.HIGHLIGHT]:
+            if solution[i - 1] not in timestep[name][Op.HIGHLIGHT]:
                 return False
             i = i - 1
     return True