+++ HISTORY.md
@@ -3,8 +3,8 @@
 ## 0.1.0
 
 ### API
-- Display dependencies on click (#25)
 - Traceback solution verification and visualization (#24)
+- Display dependencies on click (#25)
 - Migrate from graph objects to Dash (#20)
 - Enable arrow keys for time travel (#20)
 - Remove "null" text in empty cells (#21)